+++ bzt/jmx.py
@@ -23,7 +23,7 @@ from itertools import chain
 from cssselect import GenericTranslator
 
 from bzt.engine import Scenario
-from bzt.six import etree, iteritems, string_types, parse, text_type
+from bzt.six import etree, iteritems, string_types, parse, text_type, numeric_types
 
 
 class JMX(object):
@@ -316,6 +316,8 @@ class JMX(object):
     def __add_body_from_script(args, body, proxy):
         http_args_coll_prop = JMX._collection_prop("Arguments.arguments")
         for arg_name, arg_value in body.items():
+            if not (isinstance(arg_value, string_types) or isinstance(arg_value, numeric_types)):
+                raise ValueError('Body structure requires application/JSON header')
             try:
                 http_element_prop = JMX._element_prop(arg_name, "HTTPArgument")
             except ValueError: