+++ tests/modules/test_blazeMeterUploader.py
@@ -6,6 +6,7 @@ import time
 
 from bzt.modules.aggregator import DataPoint, KPISet
 from tests import BZTestCase, random_datapoint, __dir__
+from bzt.six import URLError
 from bzt.modules.blazemeter import BlazeMeterUploader, BlazeMeterClient, BlazeMeterClientEmul, ResultsFromBZA
 from tests.mocks import EngineEmul
 import bzt.modules.blazemeter
@@ -84,6 +85,16 @@ class TestBlazeMeterClientUnicode(BZTestCase):
         blazemeter_client.upload_file(__dir__() + "/../data/unicode_file")
         bzt.modules.blazemeter.urlopen = normal_urlopen
 
+    def test_binary_unicode_error(self):
+        client = BlazeMeterClient(logging.getLogger(''))
+        client.address = u"http://127.0.0.1:58000"
+        client.active_session_id = "ffff"
+        self.token = "faketoken"
+        with open(__dir__() + "/../data/jmeter-dist-2.13.zip", 'rb') as fds:
+            zip_content = fds.read()
+        # actually, we're testing that UnicodeDecodeError is not raised
+        self.assertRaises(URLError, client.upload_file, "jtls_and_more.zip", zip_content)
+
 
 class DummyHttpResponse(object):
     def __init__(self):
@@ -96,6 +107,7 @@ class DummyHttpResponse(object):
 
 
 def dummy_urlopen(*args, **kwargs):
+    del args, kwargs
     return DummyHttpResponse()